LPS-54176 Move MailServiceTestUtil to portal-test-internal, so this class can use the Preferences wrapper
