+++ portal-test-internal/src/com/liferay/portal/test/rule/callback/SynchronousMailTestCallback.java
@@ -12,10 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.kernel.test.rule.callback;
+package com.liferay.portal.test.rule.callback;
 
 import com.liferay.portal.kernel.messaging.DestinationNames;
-import com.liferay.portal.kernel.test.util.MailServiceTestUtil;
+import com.liferay.portal.kernel.test.rule.callback.SynchronousDestinationTestCallback;
+import com.liferay.portal.util.test.MailServiceTestUtil;
 
 import org.junit.runner.Description;
